+++ src/gemseo/algos/opt/scipy_linprog/scipy_linprog.py
@@ -71,7 +71,7 @@ class ScipyLinProgAlgorithmDescription(OptimizationAlgorithmDescription):
     for_linear_problems: bool = True
     """Whether the optimization algorithm is dedicated to linear problems."""
 
-    settings: type[BaseSciPyLinProgSettings] = BaseSciPyLinProgSettings
+    Settings: type[BaseSciPyLinProgSettings] = BaseSciPyLinProgSettings
     """The option validation model for SciPy linear programming library."""
 
 
@@ -120,14 +120,14 @@ class ScipyLinprog(BaseOptimizationLibrary):
             description=("Linear programming using the HiGHS interior point solver."),
             internal_algorithm_name="highs-ipm",
             website=f"{__DOC}optimize.linprog-highs-ipm.html",
-            settings=HiGHSInteriorPointSettings,
+            Settings=HiGHSInteriorPointSettings,
         ),
         "HIGHS_DUAL_SIMPLEX": ScipyLinProgAlgorithmDescription(
             algorithm_name="Dual simplex",
             description=("Linear programming using the HiGHS dual simplex solver."),
             internal_algorithm_name="highs-ds",
             website=f"{__DOC}optimize.linprog-highs-ds.html",
-            settings=HiGHSDualSimplexSettings,
+            Settings=HiGHSDualSimplexSettings,
         ),
         "HIGHS": ScipyLinProgAlgorithmDescription(
             algorithm_name="HiGHS",